+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Failed Login now writes the error code in the log
+  [#2424](https://github.com/cyberark/conjur/pull/2424)
+- Bump cyberark base images from 1.0.5 to 1.0.6
+  [#2420](https://github.com/cyberark/conjur/pull/2420)
+- Bump cyberark base images from 1.0.4 to 1.0.5
+  [#2418](https://github.com/cyberark/conjur/pull/2418)
+
+### Fixed
+- Return 401 instead of 500 for invalid basic auth header. 
+  [#1990](https://github.com/cyberark/conjur/issues/1990)
+- Added check to stop hosts from setting passwords 
+  [#1920](https://github/cyberark/conjur/issues/1920)
+
+## [1.14.1] - 2021-11-05
+
+### Fixed
+- Version bump to resolve build error on tagged master. [#2416](https://github.com/cyberark/conjur/pull/2416)
+
 ## [1.14.0] - 2021-11-03
 
 ### Added
@@ -726,7 +745,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - The first tagged version.
 
-[Unreleased]: https://github.com/cyberark/conjur/compare/v1.14.0...HEAD
+[Unreleased]: https://github.com/cyberark/conjur/compare/v1.14.1...HEAD
+[1.14.1]: https://github.com/cyberark/conjur/compare/v1.14.0...v1.14.1
 [1.14.0]: https://github.com/cyberark/conjur/compare/v1.13.2...v1.14.0
 [1.13.2]: https://github.com/cyberark/conjur/compare/v1.13.1...v1.13.2
 [1.13.1]: https://github.com/cyberark/conjur/compare/v1.13.0...v1.13.1